the directory structure modified
